02-mcp-rag/rag-server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "faiss-cpu>=1.10.0",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
]
